seed from createRun
