+++ app/src/organisms/Desktop/ChooseRobotToRunProtocolSlideout/useCreateRunFromProtocol.ts
@@ -1,9 +1,7 @@
 import { useTranslation } from 'react-i18next'
-import { useQueryClient } from 'react-query'
 import { useSelector } from 'react-redux'
 
 import {
-  getQueryKey,
   useCreateProtocolMutation,
   useCreateRunMutation,
   useHost,
@@ -45,7 +43,6 @@ export function useCreateRunFromProtocol(
   const contextHost = useHost()
   const host =
     hostOverride != null ? { ...contextHost, ...hostOverride } : contextHost
-  const queryClient = useQueryClient()
   const { t } = useTranslation('shared')
 
   const customLabwareFiles = useSelector((state: State) =>
@@ -68,19 +65,6 @@ export function useCreateRunFromProtocol(
     documentationState,
     {
       ...options,
-      onSuccess: (...args) => {
-        const [response] = args
-        queryClient.setQueryData(
-          getQueryKey(host, 'runs', response.data.id, 'details'),
-          response
-        )
-        queryClient
-          .invalidateQueries(getQueryKey(host, 'runs', 'details'))
-          .catch((e: Error) => {
-            console.error(`error invalidating runs query: ${e.message}`)
-          })
-        options.onSuccess?.(...args)
-      },
       onError: (error, variables, context) => {
         clearDocreport()
         options.onError?.(error, variables, context)
