+++ app/src/organisms/ODD/ProtocolSetup/ProtocolSetupParameters/ProtocolSetupParameters.tsx
@@ -1,6 +1,5 @@
 import { Fragment, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useQueryClient } from 'react-query'
 import { useNavigate } from 'react-router-dom'
 
 import {
@@ -10,7 +9,6 @@ import {
   SPACING,
 } from '@opentrons/components'
 import {
-  getQueryKey,
   isDocumentedMutationError,
   useCreateProtocolAnalysisMutation,
   useCreateRunMutation,
@@ -62,7 +60,6 @@ export function ProtocolSetupParameters({
   const { t } = useTranslation('protocol_setup')
   const navigate = useNavigate()
   const host = useHost()
-  const queryClient = useQueryClient()
   const [chooseValueScreen, setChooseValueScreen] =
     useState<ChoiceParameter | null>(null)
   const [showNumericalInputScreen, setShowNumericalInputScreen] =
@@ -189,17 +186,6 @@ export function ProtocolSetupParameters({
   const { createRun, isLoading: isRunLoading } = useCreateRunMutation(
     documentationState,
     {
-      onSuccess: data => {
-        queryClient.setQueryData(
-          getQueryKey(host, 'runs', data.data.id, 'details'),
-          data
-        )
-        queryClient
-          .invalidateQueries(getQueryKey(host, 'runs', 'details'))
-          .catch((e: Error) => {
-            console.error(`could not invalidate runs cache: ${e.message}`)
-          })
-      },
       onError: error => {
         if (isDocumentedMutationError(error)) {
           setStartSetup(false)
