+++ app/src/organisms/ODD/QuickTransferFlow/SummaryAndSettings.tsx
@@ -1,6 +1,5 @@
 import { useReducer, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useQueryClient } from 'react-query'
 import { useNavigate } from 'react-router-dom'
 
 import {
@@ -13,7 +12,6 @@ import {
   Tabs,
 } from '@opentrons/components'
 import {
-  getQueryKey,
   useCreateProtocolMutation,
   useCreateRunMutation,
   useHost,
@@ -55,7 +53,6 @@ export function SummaryAndSettings(
   const { exitButtonProps, state: wizardFlowState, analyticsStartTime } = props
   const navigate = useNavigate()
   const { trackEventWithRobotSerial } = useTrackEventWithRobotSerial()
-  const queryClient = useQueryClient()
   const host = useHost()
   const { t } = useTranslation(['quick_transfer', 'shared'])
   const [showSaveOrRunModal, setShowSaveOrRunModal] = useState<boolean>(false)
@@ -82,15 +79,6 @@ export function SummaryAndSettings(
     documentationState,
     {
       onSuccess: data => {
-        queryClient.setQueryData(
-          getQueryKey(host, 'runs', data.data.id, 'details'),
-          data
-        )
-        queryClient
-          .invalidateQueries(getQueryKey(host, 'runs', 'details'))
-          .catch((e: Error) => {
-            console.error(`error invalidating runs query: ${e.message}`)
-          })
         navigate(`/runs/${data.data.id}/setup`)
       },
     },
