+++ app/src/pages/ODD/ProtocolDetails/index.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useQueryClient } from 'react-query'
 import { useDispatch, useSelector } from 'react-redux'
 import { useNavigate, useParams } from 'react-router-dom'
 import last from 'lodash/last'
@@ -27,7 +26,6 @@ import {
   TYPOGRAPHY,
 } from '@opentrons/components'
 import {
-  getQueryKey,
   isDocumentedMutationError,
   useCreateRunMutation,
   useDeleteProtocolMutation,
@@ -327,7 +325,6 @@ export function ProtocolDetails(): JSX.Element | null {
   const host = useHost()
   const { makeSnackbar } = useToaster()
   const [showParameters, setShowParameters] = useState<boolean>(false)
-  const queryClient = useQueryClient()
   const [currentOption, setCurrentOption] = useState<TabOption>(
     protocolSectionTabOptions[0]
   )
@@ -354,19 +351,7 @@ export function ProtocolDetails(): JSX.Element | null {
   const { deleteProtocol } = useDeleteProtocolMutation(deleteDocumentationState)
   const { deleteRun } = useDeleteRunMutation(deleteDocumentationState)
   const documentationState = useDocumentationState()
-  const { createRun } = useCreateRunMutation(documentationState, {
-    onSuccess: data => {
-      queryClient.setQueryData(
-        getQueryKey(host, 'runs', data.data.id, 'details'),
-        data
-      )
-      queryClient
-        .invalidateQueries(getQueryKey(host, 'runs', 'details'))
-        .catch((e: Error) => {
-          console.error(`could not invalidate runs cache: ${e.message}`)
-        })
-    },
-  })
+  const { createRun } = useCreateRunMutation(documentationState)
 
   const isRobotOutOfStorage = useIsRobotOutOfStorage()
   const [showRobotOutOfStorageModal, setShowRobotOutOfStorageModal] =
