+++ app/src/pages/ODD/QuickTransferDetails/index.tsx
@@ -1,6 +1,5 @@
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useQueryClient } from 'react-query'
 import { useDispatch, useSelector } from 'react-redux'
 import { useNavigate, useParams } from 'react-router-dom'
 
@@ -25,9 +24,7 @@ import {
   TYPOGRAPHY,
 } from '@opentrons/components'
 import {
-  getQueryKey,
   useCreateRunMutation,
-  useHost,
   useProtocolQuery,
 } from '@opentrons/react-api-client'
 
@@ -308,9 +305,7 @@ export function QuickTransferDetails(): JSX.Element | null {
   )
 
   const dispatch = useDispatch<Dispatch>()
-  const host = useHost()
   const { makeSnackbar } = useToaster()
-  const queryClient = useQueryClient()
   const [currentOption, setCurrentOption] = useState<TabOption>(
     transferSectionTabOptions[0]
   )
@@ -328,19 +323,7 @@ export function QuickTransferDetails(): JSX.Element | null {
   const pinned = pinnedTransferIds.includes(transferId)
   const documentationState = useDocumentationState()
 
-  const { createRun } = useCreateRunMutation(documentationState, {
-    onSuccess: data => {
-      queryClient.setQueryData(
-        getQueryKey(host, 'runs', data.data.id, 'details'),
-        data
-      )
-      queryClient
-        .invalidateQueries(getQueryKey(host, 'runs', 'details'))
-        .catch((e: Error) => {
-          console.error(`could not invalidate runs cache: ${e.message}`)
-        })
-    },
-  })
+  const { createRun } = useCreateRunMutation(documentationState)
 
   const handlePinClick = (): void => {
     if (!pinned) {
